fix masonry
