+++ src/components/loaders/last-posts-loader.tsx
@@ -1,0 +1,14 @@
+import Posts from "@/components/posts"
+import Loader from "@/components/common/loader"
+import useLastPosts from "@/hooks/use-last-posts"
+
+export const LastPostsLoader = () => {
+  const posts = useLastPosts()
+
+  if (!posts) return <Loader size="lg" />
+  if (!posts.length) return <div>Aucune publication pour le moment...</div>
+
+  return <Posts posts={posts} />
+}
+
+export default LastPostsLoader
